requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
SQLAlchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
passlib[bcrypt]==1.7.4
pydantic-settings==2.4.0
python-dotenv==1.0.1
email-validator==2.2.0
httpx==0.25.2
bcrypt==4.0.1
jinja2==3.1.4
